+++ documentation/changelog.md
@@ -29,6 +29,7 @@ description: List of all the changes and current versions.
 + Added Storage System
 + Added Storage Collection
 + Added Storage Containers for Interiors
++ Added Vehicle Storage Trunks
 ```
 
 ## 2.0.2